+++ src/main/java/ru/native_speakers/cinema_expert_api/service/ImagesService.java
@@ -5,7 +5,7 @@ import ru.native_speakers.cinema_expert_api.model.Image;
 import java.util.List;
 
 public interface ImagesService {
-    Image findImageByImageId(int imageId);
+    Image findImageByImageId(long imageId);
     List<Image> findImages(int pageSize, int page);
     long getImagesCount();
 }
